requirements.txt
asgiref==3.3.4
Django==3.2.4
django-environ==0.4.5
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
django-search-admin-autocomplete
django-filebrowser
django-grappelli
whitenoise
django-cors-headers==3.7.0
Pillow

#debugging package
django-extensions 
ipython 
sqlite-connector

#fast form 
django-crispy-forms